+++ src/web.mjs
@@ -9,13 +9,68 @@ export * from "./core.mjs";
 
 /**
  * @func fetchStrict
- * @desc Same as `fetch()` but rejects if HTTP error (e.g "404 not found")
+ * @desc Similar to `fetch()` but
+ *  1. rejects if HTTP error (e.g "404 not found").
+ *  2. if `body` is a normal object (such as those created by `{}`), convert it into `URLSearchParams`.
+ *  3. if `method` is set to 'GET' explicitly,
+ *     and `body` is `URLSearchParams` or a normal object,
+ *     convert the request into an URL with search param.
+ *  4. if `body` exists without setting `method`, warn and request with POST method.
  * @param {string | URL | Request} resource
- * @param {Object} options - same as `fetch()`
+ * @param {Object} options
  * @returns {Promise.<Response>}
  */
-export async function fetchStrict(...args) {
-    const response = await fetch(...args);
+export async function fetchStrict(resource, options) {
+    let body = resource.body || options?.body;
+    if (body) {
+        if (body.constructor === Object) {
+            body = new URLSearchParams(body);
+            if (resource.body) resource = new Request(resource, {body});
+            if (options.body) options.body = body;
+        }
+
+        const method = resource.method || options?.method;
+        switch (method) {
+            case 'GET': {
+                if (body instanceof URLSearchParams) {
+                    let url = resource.url ?? (resource + '');
+                    try {
+                        url = new URL(url); // full url
+                    }
+                    catch (err) { // path only
+                        if (url.startsWith('/')) { // abs path
+                            url = (new URL(document.baseURI)).origin + url;
+                        }
+                        else { // relative path
+                            const pos = document.baseURI.lastIndexOf('/');
+                            url = document.baseURI.slice(0, pos + 1) + url;
+                        }
+                        url = new URL(url);
+                    }
+                    // const url = new URL(resource.url ?? (resource + ''));
+                    for (const [key, value] of body.entries())
+                        url.searchParams.set(key, value);
+                    resource = resource.url ? (new Request(resource, {url, body: undefined})) : url;
+                    if (options?.body) delete options.body;
+                }
+                break;
+            }
+            case 'POST':
+            case 'PUT': {
+                break;
+            }
+            default: {
+                console.warn(
+                    'While using `fetch()` with `body`, `method` shall be set to "POST" or "PUT".'
+                    + ` You've wrongly set it to "${method}".`
+                );
+                if (resource.url) resource = new Request(resource, {method: 'POST'});
+                if (typeof options === 'object') options.method = 'POST';
+            }
+        }
+    }
+
+    const response = await fetch(resource, options);
     if (response.ok) return response;
     throw new ReferenceError(response.statusText);
 }
@@ -80,13 +135,18 @@ export function readFile(blob, type) {
 /**
  * @func createFormData
  * @desc Creates a new `FormData` from an object
- * @param {Object} object
+ * @param {Object | URLSearchParams} object
  * @returns FormData
  */
 export function createFormData(object) {
     const fd = new FormData();
-    for (const key in object)
-        fd.append(key, object[key]);
+
+    if (object instanceof URLSearchParams) {
+        for (const [key, value] of object.entries())
+            fd.append(key, value);
+    }
+    else for (const key in object) fd.append(key, object[key]);
+
     return fd;
 }